apps/glue-python/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-dotenv>=1.0
requests>=2.31
pydantic>=2.6

pandas>=2.2
openpyxl>=3.1
python-docx>=1.1
python-pptx>=1.0

polars>=0.20
pyyaml>=6.0
pypdf>=4.2
Pillow>=10.3
pytesseract>=0.3.10
